Ensure all function arguments are JavaScript objects
This is pyodide specific but should be a noop for plain JavaScript
consumers. In pyodide, some function arguments might be Proxy objects
so we first need to convert them to JavaScript objects before use.

We don’t need to apply this to everything (e.g. strings, booleans) but
I’m worried about missing some argument, e.g. if something is normally
a string but Highcharts also allows it to be an array.