Define getters and setters for chart properties
In commit 5666fce, I mistakenly thought that the $loc.render method
wasn’t doing anything but now I realise it was updating this._options
with any top-level attributes that had been set on the chart in Python:

```python
bar_chart = pygal.Bar(title="Foo")
bar_chart.title = "Bar"
```

In this ^ case, it was fetching the ‘title’ property from the bar_chart
instance and setting it in the _options object so that it can be used
for rendering.

We could repeat this pattern at the top of the Chart#render method but
I think it’s more idiomatic to implement getters and setters for all
chart properties so that this._options is always up to date, without
baving to call Chart#render first. For example, someone might implement
another method that depends on title and calls it before #render.